Move stats package to pkg/stats
