+++ pkg/stats/timer.go
