+++ promql/engine.go
@@ -25,7 +25,7 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
-	"github.com/prometheus/prometheus/stats"
+	"github.com/prometheus/prometheus/pkg/stats"
 	"github.com/prometheus/prometheus/storage/local"
 	"github.com/prometheus/prometheus/storage/metric"
 )
